Tendermint update 0.27.4 (#333)
* Update tendermint & jtendermint

* Update tendermint config case classes

* Increased timeout
